Started refactoring, and Created observables for Facebook Sdk
As the title states, more refactoring needs to be done however
