+++ client/secureChatIonic/app/providers/app-messaging/app-messaging.ts
@@ -5,12 +5,6 @@ import 'rxjs/add/operator/map';
 //Import our providers
 import { AppSettings } from '../../providers/app-settings/app-settings';
 
-/*
-  Generated class for the AppMessaging provider.
-
-  See https://angular.io/docs/ts/latest/guide/dependency-injection.html
-  for more info on providers and Angular 2 DI.
-*/
 @Injectable()
 export class AppMessaging {
