+++ client/secureChatIonic/app/theme/app.core.scss
@@ -9,7 +9,7 @@
 
 //Import our page scss
 @import "../pages/home/home";
-@import "../pages/all-messages/all-messages";
+@import "../pages/all-conversations/all-conversations";
 @import "../pages/conversation/conversation";
 @import "../pages/auth-login/auth-login";
